Initailly restructure test cases for #625
