+++ tests/src/test/java/org/wso2/micro/gateway/tests/http2/HTTP2RequestsWithHTTP1BackEndTestCase.java
@@ -16,12 +16,8 @@
  *  under the License.
  */
 
-package org.wso2.micro.gateway.tests.services;
+package org.wso2.micro.gateway.tests.http2;
 
-import io.netty.handler.codec.http.HttpHeaderNames;
-import io.netty.handler.codec.http.HttpHeaderValues;
-import io.netty.handler.codec.http.HttpScheme;
-import io.netty.handler.codec.http2.HttpConversionUtil;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.testng.Assert;
@@ -30,7 +26,6 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import org.wso2.micro.gateway.tests.common.BaseTestCase;
 import org.wso2.micro.gateway.tests.common.CLIExecutor;
-import org.wso2.micro.gateway.tests.common.HTTP2Server.MockHttp2Server;
 import org.wso2.micro.gateway.tests.common.KeyValidationInfo;
 import org.wso2.micro.gateway.tests.common.MockAPIPublisher;
 import org.wso2.micro.gateway.tests.common.MockHttpServer;
@@ -39,21 +34,16 @@ import org.wso2.micro.gateway.tests.common.model.ApplicationDTO;
 import org.wso2.micro.gateway.tests.context.ServerInstance;
 import org.wso2.micro.gateway.tests.context.Utils;
 import org.wso2.micro.gateway.tests.util.HTTP2Client.Http2ClientRequest;
-import org.wso2.micro.gateway.tests.util.HttpClientRequest;
 import org.wso2.micro.gateway.tests.util.TestConstant;
 
 import java.io.File;
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.wso2.micro.gateway.tests.util.TestConstant.GATEWAY_LISTENER_HTTPS_PORT;
 import static org.wso2.micro.gateway.tests.util.TestConstant.GATEWAY_LISTENER_HTTP_PORT;
 
-public class HTTP2RequestsWithHTTP2BackEndTestCase extends BaseTestCase {
+public class HTTP2RequestsWithHTTP1BackEndTestCase extends BaseTestCase {
 
-    protected final static int MOCK_HTTP2_SERVER_PORT = 8443;
     private static final Log log = LogFactory.getLog(HTTP2RequestsWithHTTP2BackEndTestCase.class);
-    protected MockHttp2Server mockHttp2Server;
     protected Http2ClientRequest http2ClientRequest;
     private String jwtTokenProd;
 
@@ -66,7 +56,7 @@ public class HTTP2RequestsWithHTTP2BackEndTestCase extends BaseTestCase {
         API api = new API();
         api.setName("PizzaShackAPI");
         api.setContext("/pizzashack");
-        api.setEndpoint("https://localhost:8443");
+        api.setEndpoint(getMockServiceURLHttp("/echo"));
         api.setVersion("1.0.0");
         api.setProvider("admin");
         //Register API with label
@@ -105,53 +95,22 @@ public class HTTP2RequestsWithHTTP2BackEndTestCase extends BaseTestCase {
         String balPath = CLIExecutor.getInstance().getLabelBalx(project);
         String configPath = getClass().getClassLoader()
                 .getResource("confs" + File.separator + "http2-test.conf").getPath();
-        String[] args = {"--config", configPath};
+        String[] args = {"--config", configPath, "--experimental"};
         microGWServer.startMicroGwServer(balPath, args);
 
         jwtTokenProd = getJWT(api, application, "Unlimited", TestConstant.KEY_TYPE_PRODUCTION, 3600);
-
-        boolean isOpen2 = Utils.isPortOpen(MOCK_HTTP2_SERVER_PORT);
-        Assert.assertFalse(isOpen2, "Port: " + MOCK_HTTP2_SERVER_PORT + " already in use.");
-        mockHttp2Server = new MockHttp2Server(MOCK_HTTP2_SERVER_PORT, true);
-        mockHttp2Server.start();
     }
 
-    @Test(description = "Test API invocation with an HTTP/2.0 request via insecure connection sending to HTTP/2.0 BE")
-    public void testHTTP2RequestsViaInsecureConnectionWithHTTP2BE() throws Exception {
+    @Test(description = "Test API invocation with an HTTP/2.0 request via insecure connection sending to HTTP/1.1 BE")
+    public void testHTTP2RequestsViaInsecureConnectionWithHTTP1BE() throws Exception {
         http2ClientRequest = new Http2ClientRequest(false, GATEWAY_LISTENER_HTTP_PORT, jwtTokenProd);
         http2ClientRequest.start();
     }
 
-    @Test(description = "Test API invocation with an HTTP/2.0 request via secure connection sending to HTTP/2.0 BE")
-    public void testHTTP2RequestsViaSecureConnectionWithHTTP2BE() throws Exception {
+    @Test(description = "Test API invocation with an HTTP/2.0 request via secure connection sending to HTTP/1.1 BE")
+    public void testHTTP2RequestsViaSecureConnectionWithHTTP1BE() throws Exception {
         http2ClientRequest = new Http2ClientRequest(true, GATEWAY_LISTENER_HTTPS_PORT, jwtTokenProd);
         http2ClientRequest.start();
-    }
-
-    @Test(description = "Test API invocation with an HTTP/1.1 request via insecure connection sending to HTTP/2.0 BE")
-    public void testHTTP1RequestsViaInsecureConnectionWithHTTP2BE() throws Exception {
-        Map<String, String> headers = new HashMap<>();
-        headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtTokenProd);
-        headers.put(HttpHeaderNames.HOST.toString(), "127.0.0.1:9590");
-        headers.put(HttpConversionUtil.ExtensionHeaderNames.SCHEME.text().toString(), HttpScheme.HTTP.toString());
-        headers.put(HttpHeaderNames.ACCEPT_ENCODING.toString(), HttpHeaderValues.GZIP.toString());
-        headers.put(HttpHeaderNames.ACCEPT_ENCODING.toString(), HttpHeaderValues.DEFLATE.toString());
-        org.wso2.micro.gateway.tests.util.HttpResponse response = HttpClientRequest
-                .doGet(getServiceURLHttp("/pizzashack/1.0.0/menu"), headers);
-        log.info("Response: " + response.getResponseMessage() + " , " + response.getResponseCode());
-    }
-
-    @Test(description = "Test API invocation with an HTTP/1.1 request via secure connection sending to HTTP/2.0 BE")
-    public void testHTTP1RequestsViaSecureConnectionWithHTTP2BE() throws Exception {
-        Map<String, String> headers = new HashMap<>();
-        headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtTokenProd);
-        headers.put(HttpHeaderNames.HOST.toString(), "127.0.0.1:9595");
-        headers.put(HttpConversionUtil.ExtensionHeaderNames.SCHEME.text().toString(), HttpScheme.HTTP.toString());
-        headers.put(HttpHeaderNames.ACCEPT_ENCODING.toString(), HttpHeaderValues.GZIP.toString());
-        headers.put(HttpHeaderNames.ACCEPT_ENCODING.toString(), HttpHeaderValues.DEFLATE.toString());
-        org.wso2.micro.gateway.tests.util.HttpResponse response = HttpClientRequest
-                .doGet(getServiceURLHttp("/pizzashack/1.0.0/menu"), headers);
-        log.info("Response: " + response.getResponseMessage() + " , " + response.getResponseCode());
     }
 
     @AfterClass
