+++ tests/src/test/java/org/wso2/micro/gateway/tests/validation/ValidationTestCase.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.micro.gateway.tests.toolkit;
+package org.wso2.micro.gateway.tests.validation;
 
 import io.netty.handler.codec.http.HttpHeaderNames;
 import org.testng.Assert;
