+++ tests/src/test/resources/testng.xml
@@ -29,11 +29,11 @@
     <test name="micro-gw-services" preserve-order="true" parallel="false">
         <classes>
             <class name="org.wso2.micro.gateway.tests.jwtRevocation.JWTRevocationSupportTestCase" />
-            <class name="org.wso2.micro.gateway.tests.services.AuthenticationFailureTestCase" />
-            <class name="org.wso2.micro.gateway.tests.services.MutualSSLTestCase" />
-            <class name="org.wso2.micro.gateway.tests.services.HTTP2RequestsWithHTTP1BackEndTestCase" />
-            <class name="org.wso2.micro.gateway.tests.services.HTTP2RequestsWithHTTP2BackEndTestCase" />
-            <class name="org.wso2.micro.gateway.tests.services.PreRequisites" />
+            <class name="org.wso2.micro.gateway.tests.security.AuthenticationFailureTestCase" />
+            <class name="org.wso2.micro.gateway.tests.security.MutualSSLTestCase" />
+            <class name="org.wso2.micro.gateway.tests.http2.HTTP2RequestsWithHTTP1BackEndTestCase" />
+            <class name="org.wso2.micro.gateway.tests.http2.HTTP2RequestsWithHTTP2BackEndTestCase" />
+            <class name="org.wso2.micro.gateway.tests.throttling.PreRequisites" />
 
         </classes>
     </test>
@@ -41,16 +41,16 @@
     <test name="micro-gw-toolkit" preserve-order="true" parallel="false">
         <classes>
             <!-- DistributedThrottlingTestCase should be the first since it starts message broker -->
-            <class name="org.wso2.micro.gateway.tests.toolkit.DistributedThrottlingTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.throttling.DistributedThrottlingTestCase"/>
 
-            <class name="org.wso2.micro.gateway.tests.toolkit.APIInvokeWithOAuth2andBasicAuthTestCase"/>
-            <class name="org.wso2.micro.gateway.tests.toolkit.APIInvokeWithBasicAuthTestCase"/>
-            <class name="org.wso2.micro.gateway.tests.toolkit.APIInvokeWithOAuthTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.security.APIInvokeWithOAuth2andBasicAuthTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.security.APIInvokeWithBasicAuthTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.security.APIInvokeWithOAuthTestCase"/>
 
-            <class name="org.wso2.micro.gateway.tests.toolkit.ThrottlingTestCase"/>
-            <class name="org.wso2.micro.gateway.tests.toolkit.OASAPIInvokeTestCase"/>
-            <class name="org.wso2.micro.gateway.tests.toolkit.CookieAuthTestCase"/>
-            <class name="org.wso2.micro.gateway.tests.toolkit.ValidationTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.throttling.ThrottlingTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.extensions.OASAPIInvokeTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.security.CookieAuthTestCase"/>
+            <class name="org.wso2.micro.gateway.tests.validation.ValidationTestCase"/>
         </classes>
     </test>
 </suite>
